+++ code/web/public/index.html
@@ -25,8 +25,7 @@
     <noscript>
       You need to enable JavaScript to run this app.
     </noscript>
-    <div id="login"></div>
-    <div id="feed"></div>
+    <div id="root"></div>
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
